Replace Tree-First Model with ID Model
Instead of building a tree of items and maintaining a map of routes to items,
build maps of routes to ids, ids to items, and maintain parent/child links
for convenience.

Distinct VirtualItems have been removed and replaced with a way to associate
virtual data with IDs, with a way to add virtual data to routes before a page
exists at that route.